requirements.txt
fastapi[standard]
uvicorn[standard]
llama-index-llms-ollama
llama-index-vector-stores-chroma
llama-index-vector-stores-qdrant
llama-index-core
llama-index-readers-file
llama-index-llms-ollama
llama-index-embeddings-huggingface
llama-index-embeddings-nomic
llama-index-embeddings-ollama
ollama
PyYAML
chromadb
torch==2.3.1
python-dotenv
pydantic
streamlit